+++ src/main/java/chess/controller/mapper/response/board/BlackPieceSymbolMapper.java
@@ -1,4 +1,4 @@
-package chess.controller.mapper.response;
+package chess.controller.mapper.response.board;
 
 import chess.domain.piece.Bishop;
 import chess.domain.piece.King;
@@ -10,25 +10,25 @@ import chess.domain.piece.Rook;
 import java.util.Arrays;
 import java.util.NoSuchElementException;
 
-public enum WhitePieceSymbolMapper {
-    PAWN(Pawn.class, "p"),
-    BISHOP(Bishop.class, "b"),
-    KING(King.class, "k"),
-    KNIGHT(Knight.class, "n"),
-    QUEEN(Queen.class, "q"),
-    ROOK(Rook.class, "r"),
+public enum BlackPieceSymbolMapper {
+    PAWN(Pawn.class, "P"),
+    BISHOP(Bishop.class, "B"),
+    KING(King.class, "K"),
+    KNIGHT(Knight.class, "N"),
+    QUEEN(Queen.class, "Q"),
+    ROOK(Rook.class, "R"),
     ;
 
     private final Class<? extends Piece> piece;
     private final String viewSymbol;
 
-    WhitePieceSymbolMapper(Class<? extends Piece> piece, String viewSymbol) {
+    BlackPieceSymbolMapper(Class<? extends Piece> piece, String viewSymbol) {
         this.piece = piece;
         this.viewSymbol = viewSymbol;
     }
 
     public static String getViewSymbolBy(Piece piece) {
-        return Arrays.stream(WhitePieceSymbolMapper.values())
+        return Arrays.stream(BlackPieceSymbolMapper.values())
                 .filter(it -> it.piece.isInstance(piece))
                 .findFirst()
                 .orElseThrow(NoSuchElementException::new)
